moved f-ns addModal & closeModal
