+++ src/components/modal/modal.js
@@ -1,0 +1,33 @@
+import { getCardDatabById } from "../../api/getApi";
+import { addMenu } from "../card/createCards";
+
+export function addModal(event) {
+  if(!event.target.classList.contains('card__btn')) {
+    const cardId = +event.currentTarget.id;
+    const overlay = document.querySelector('.overlay');
+    const body = document.querySelector('body');
+
+    try {
+      getCardDatabById(cardId)
+      .then(cardData => { 
+          addMenu(cardData, 'modal');
+          overlay.style.display = 'block';
+          overlay.addEventListener('click', closeModal);
+          body.classList.add('modal-open');
+      });
+      
+    } catch(error) {
+      throw error;
+    }
+  }
+}
+
+export function closeModal() {
+  const card = document.querySelector('.card_modal');
+  const overlay = document.querySelector('.overlay');
+  const body = document.querySelector('body');
+
+  card.remove();
+  overlay.style.display = 'none';
+  body.classList.remove('modal-open');
+}
