+++ src/pages/orders.html
@@ -67,7 +67,7 @@
         </div>
     </footer>
 
-    <script src="/src/components/orders/order.js"></script>
+    <script src="/src/components/orders/createOrderElements.js"></script>
 
 </body>
 </html>
