+++ src/styles/pages/orders/orders.scss
@@ -5,6 +5,7 @@
         margin: 0 auto;
         display: flex;
         flex-direction: column;
+        min-height: 75vh;
     }
 
     &__title {
